refactor(workflow): reorganize feature extraction nodes into dim1d/dim2d subdirectories
- Move existing nodes (MelSpectrogram, MFCC, STFT, FFT, Statistics) to categorized subdirectories
- Add CQTNode for constant-Q transform feature extraction
- Add SpectralContrastNode for spectral contrast features
- Add PitchNode for fundamental frequency tracking
- Add SpectralFlatnessNode for tonality detection
- Add extract_pitch() and extract_spectral_flatness() to FeatureExtractor
- Create feature/base.py with shared FeatureData and utilities
