+++ src/workflow/nodes/feature/dim1d/__init__.py
@@ -1,0 +1,18 @@
+# -*- coding: utf-8 -*-
+"""
+1D Feature Extraction Nodes
+
+One-dimensional feature extraction nodes suitable for 1D CNN or fully connected networks.
+"""
+
+from .fft import FFTNode
+from .statistics import StatisticsNode
+from .pitch import PitchNode
+from .spectral_flatness import SpectralFlatnessNode
+
+__all__ = [
+    'FFTNode',
+    'StatisticsNode',
+    'PitchNode',
+    'SpectralFlatnessNode',
+]
