+++ src/workflow/nodes/feature/dim1d/statistics.py
@@ -1,0 +1,131 @@
+# -*- coding: utf-8 -*-
+"""
+Statistics Feature Extraction Node
+
+Extracts spectral statistical features (1D) from audio signals.
+"""
+
+import librosa
+import numpy as np
+
+from ....node_base import BaseNode, NodeCategory, register_node
+from ....port import DataType
+from ..base import FeatureData, extract_from_audio_or_list, validate_audio_input, AudioData
+
+
+@register_node
+class StatisticsNode(BaseNode):
+    """
+    统计特征节点
+    
+    提取频谱统计特征（1D）。
+    """
+    
+    node_type = "statistics"
+    display_name = "统计特征"
+    category = NodeCategory.FEATURE
+    description = "提取频谱统计特征"
+    icon = "📉"
+    
+    def _setup_ports(self):
+        self.add_input("audio", DataType.AUDIO, "音频")
+        self.add_output("feature", DataType.FEATURE, "统计特征")
+    
+    def _setup_parameters(self):
+        self.add_parameter(
+            "spectral_centroid", "bool", True,
+            display_name="频谱质心"
+        )
+        self.add_parameter(
+            "spectral_bandwidth", "bool", True,
+            display_name="频谱带宽"
+        )
+        self.add_parameter(
+            "spectral_rolloff", "bool", True,
+            display_name="频谱滚降"
+        )
+        self.add_parameter(
+            "zero_crossing_rate", "bool", True,
+            display_name="过零率"
+        )
+        self.add_parameter(
+            "rms", "bool", True,
+            display_name="RMS能量"
+        )
+        self.add_parameter(
+            "aggregate", "choice", "mean",
+            display_name="聚合方式",
+            choices=["mean", "std", "mean_std", "all"]
+        )
+    
+    def execute(self) -> bool:
+        try:
+            audio_input = validate_audio_input(
+                self.get_input_data("audio"),
+                self.display_name
+            )
+        except (ValueError, TypeError) as e:
+            self.error_message = str(e)
+            return False
+        
+        aggregate = self.get_parameter("aggregate")
+        
+        def aggregate_feature(feat: np.ndarray) -> np.ndarray:
+            """Aggregate single channel feature, feat shape is (1, frames)"""
+            if aggregate == "mean":
+                return np.mean(feat, axis=1)
+            elif aggregate == "std":
+                return np.std(feat, axis=1)
+            elif aggregate == "mean_std":
+                return np.concatenate([np.mean(feat, axis=1), np.std(feat, axis=1)])
+            else:  # all - return complete feature
+                return feat.flatten()
+        
+        def extract_stats(audio: AudioData) -> FeatureData:
+            # Extract statistics for each channel
+            channel_features = []
+            
+            for ch in range(audio.channels):
+                ch_data = audio.get_channel(ch)
+                features = []
+                
+                if self.get_parameter("spectral_centroid"):
+                    feat = librosa.feature.spectral_centroid(y=ch_data, sr=audio.sample_rate)
+                    features.append(aggregate_feature(feat))
+                
+                if self.get_parameter("spectral_bandwidth"):
+                    feat = librosa.feature.spectral_bandwidth(y=ch_data, sr=audio.sample_rate)
+                    features.append(aggregate_feature(feat))
+                
+                if self.get_parameter("spectral_rolloff"):
+                    feat = librosa.feature.spectral_rolloff(y=ch_data, sr=audio.sample_rate)
+                    features.append(aggregate_feature(feat))
+                
+                if self.get_parameter("zero_crossing_rate"):
+                    feat = librosa.feature.zero_crossing_rate(ch_data)
+                    features.append(aggregate_feature(feat))
+                
+                if self.get_parameter("rms"):
+                    feat = librosa.feature.rms(y=ch_data)
+                    features.append(aggregate_feature(feat))
+                
+                if not features:
+                    features.append(np.zeros(1))
+                
+                combined = np.concatenate(features)
+                channel_features.append(combined)
+            
+            # Stack as (channels, n_features)
+            stacked = np.stack(channel_features, axis=0)
+            
+            return FeatureData(
+                data=stacked.astype(np.float32),
+                feature_type="statistics",
+                sample_rate=audio.sample_rate,
+                hop_length=512,
+                source_file=audio.file_path
+            )
+        
+        result = extract_from_audio_or_list(audio_input, extract_stats)
+        self.set_output_data("feature", result)
+        return True
